consumer/requirements.txt
aiokafka==0.10.0
cassandra-driver==3.29.1
python-dotenv==1.0.1